+++ Frontend/UpFluxAutomation/Abstractions/IStep.cs
@@ -1,11 +1,11 @@
 ﻿using System.Threading.Tasks;
 using Microsoft.Playwright;
 
-namespace UpFluxAutomation.Steps.Abstractions
+namespace UpFluxAutomation.Abstractions
 {
     public interface IStep
     {
-        Task Execute(IPage page);
+        Task Execute();
         IStep Chain(IStep next);
     }
 }
