+++ Frontend/UpFluxAutomation/Steps/LogIn/NavigateToLogin.cs
@@ -1,17 +1,19 @@
 ﻿using System.Threading.Tasks;
 using Microsoft.Playwright;
+using UpFluxAutomation.Abstractions;
 using UpFluxAutomation.Helpers;
-using UpFluxAutomation.Steps.Abstractions;
 
 namespace UpFluxAutomation.Steps
 {
     public class NavigateToLogin : BaseStep
     {
-        public NavigateToLogin(MemoryRepository repository) : base(repository) { }
+        public NavigateToLogin(IRepository repository, IStep next = null) : base(repository, next) { }
 
-        protected override async Task PerformExecute(IPage page)
+        protected override async Task PerformExecute()
         {
             Console.WriteLine("Navigating to Login Page...");
+
+            var page = Repository.Get<IPage>();
 
             await page.GotoAsync("/login");
