+++ Frontend/UpFluxAutomationTest/Assertion/EngineerLoginAssertion.cs
@@ -1,20 +1,23 @@
 ﻿using System.Threading.Tasks;
 using Microsoft.Playwright;
+using UpFluxAutomation.Abstractions;
 using UpFluxAutomation.Helpers;
 using UpFluxAutomation.Steps;
-using UpFluxAutomation.Steps.Abstractions;
 
 namespace UpFluxAutomationTest.Assertion
 {
     public class EngineerLoginAssertion : BaseStep
     {
-        public EngineerLoginAssertion(MemoryRepository repository) : base(repository) { }
+        public EngineerLoginAssertion(IRepository repository, IStep next = null) : base(repository, next) { }
 
-        protected override async Task PerformExecute(IPage page)
+        protected override async Task PerformExecute()
         {
             Console.WriteLine("Verifying login success message...");
 
-             var successMessageLocator = page.Locator("h4:has-text('You have been logged in successfully')");
+            var page = Repository.Get<IPage>();
+
+            await page.WaitForLoadStateAsync();
+            var successMessageLocator = page.Locator("h4:has-text('You have been logged in successfully')");
 
             // Perform the assertion
             await Assertions.Expect(successMessageLocator).ToBeVisibleAsync();
@@ -23,3 +26,4 @@ namespace UpFluxAutomationTest.Assertion
         }
     }
 }
+
